feat: null types improvements

+++ src/Application/Services/TerritoryStories/TerritoryStoryService.cs
@@ -346,8 +346,12 @@ public class TerritoryStoryService : ServiceRead<TerritoryStory, TerritoryStoryD
         if (hasDuplicatedEntities)
         {
             var like = await territoryStoryLikeRepository.GetByTerritoryStoryAndUserNameAsync(entityData.TerritoryStoryId, entityData.UserName, ct);
-            await territoryStoryLikeRepository.DeleteAsync(like!, ct);
-            logger.AddLog(LogType.Delete, $"Unliked territory story", "{@EntityData}", entityData);
+
+            if (like != null)
+            {
+                await territoryStoryLikeRepository.DeleteAsync(like!, ct);
+                logger.AddLog(LogType.Delete, $"Unliked territory story", "{@EntityData}", entityData);
+            }
         }
         else
         {
